Test SelectedEntriesModal in isolation
