+++ packages/core/admin/admin/src/content-manager/components/ListViewTable/BulkActionButtons/index.js
@@ -1,22 +1,7 @@
 import * as React from 'react';
 
-import {
-  Button,
-  Box,
-  Flex,
-  Dialog,
-  DialogBody,
-  DialogFooter,
-  Typography,
-  ModalLayout,
-  ModalHeader,
-  ModalBody,
-  ModalFooter,
-  Tbody,
-  Td,
-  Tr,
-} from '@strapi/design-system';
-import { useTracking, useTableContext, Table } from '@strapi/helper-plugin';
+import { Button, Flex, Dialog, DialogBody, DialogFooter, Typography } from '@strapi/design-system';
+import { useTracking, useTableContext } from '@strapi/helper-plugin';
 import { Check, ExclamationMarkCircle, Trash } from '@strapi/icons';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
@@ -25,7 +10,8 @@ import { useSelector } from 'react-redux';
 import { listViewDomain } from '../../../pages/ListView/selectors';
 import { getTrad } from '../../../utils';
 import InjectionZoneList from '../../InjectionZoneList';
-import { Body } from '../Body';
+
+import SelectedEntriesModal from './SelectedEntriesModal';
 
 /* -------------------------------------------------------------------------------------------------
  * ConfirmBulkActionDialog
@@ -207,145 +193,6 @@ const ConfirmDialogDeleteAll = ({ isOpen, onToggleDialog, isConfirmButtonLoading
 };
 
 ConfirmDialogDeleteAll.propTypes = confirmDialogsPropTypes;
-
-/* -------------------------------------------------------------------------------------------------
- * BoldChunk
- * -----------------------------------------------------------------------------------------------*/
-
-const BoldChunk = (chunks) => <Typography fontWeight="bold">{chunks}</Typography>;
-
-/* -------------------------------------------------------------------------------------------------
- * SelectedEntriesTableContent
- * -----------------------------------------------------------------------------------------------*/
-
-const SelectedEntriesTableContent = ({ notifySelectionChange }) => {
-  const { selectedEntries, setSelectedEntries, rows } = useTableContext();
-
-  // Get main field from list view layout
-  const listViewStore = useSelector(listViewDomain());
-  const { mainField } = listViewStore.contentType.settings;
-  const shouldDisplayMainField = mainField != null && mainField !== 'id';
-
-  // Select all entries by default
-  React.useEffect(() => {
-    setSelectedEntries(rows.map((row) => row.id));
-  }, [setSelectedEntries, rows]);
-
-  // Notify parent component when selection changes
-  React.useEffect(() => {
-    notifySelectionChange(selectedEntries);
-  }, [selectedEntries, notifySelectionChange]);
-
-  return (
-    <Table.Content>
-      <Table.Head>
-        <Table.HeaderCheckboxCell />
-        <Table.HeaderCell fieldSchemaType="number" label="id" name="id" />
-        {shouldDisplayMainField && (
-          <Table.HeaderCell fieldSchemaType="string" label="name" name="name" />
-        )}
-      </Table.Head>
-      <Tbody>
-        {rows.map((entry, index) => (
-          <Tr key={entry.id}>
-            <Body.CheckboxDataCell rowId={entry.id} index={index} />
-            <Td>
-              <Typography>{entry.id}</Typography>
-            </Td>
-            {shouldDisplayMainField && (
-              <Td>
-                <Typography>{entry[mainField]}</Typography>
-              </Td>
-            )}
-          </Tr>
-        ))}
-      </Tbody>
-    </Table.Content>
-  );
-};
-
-SelectedEntriesTableContent.propTypes = {
-  notifySelectionChange: PropTypes.func.isRequired,
-};
-
-/* -------------------------------------------------------------------------------------------------
- * SelectedEntriesModal
- * -----------------------------------------------------------------------------------------------*/
-
-const SelectedEntriesModal = ({ isOpen, onToggle, onConfirm }) => {
-  const { formatMessage } = useIntl();
-  const { rows, selectedEntries } = useTableContext();
-
-  // Only used to mirror the state of the nested table component
-  // setEntriesToPublish should not be called directly
-  const [entriesToPublish, setEntriesToPublish] = React.useState([]);
-
-  // Get the selected entries full data, and keep the list view order
-  // Memoize to prevent infinite useEffect runs in SelectedEntriesTableContent
-  const entries = React.useMemo(() => {
-    return rows.filter((row) => {
-      return selectedEntries.includes(row.id);
-    });
-  }, [rows, selectedEntries]);
-
-  if (!isOpen) {
-    return null;
-  }
-
-  return (
-    <ModalLayout onClose={onToggle} labelledBy="title">
-      <ModalHeader>
-        <Typography fontWeight="bold" textColor="neutral800" as="h2" id="title">
-          {formatMessage({
-            id: getTrad('containers.ListPage.selectedEntriesModal.title'),
-            defaultMessage: 'Publish entries',
-          })}
-        </Typography>
-      </ModalHeader>
-      <ModalBody>
-        <Typography>
-          {formatMessage(
-            {
-              id: getTrad('containers.ListPage.selectedEntriesModal.selectedCount'),
-              defaultMessage:
-                '<b>{count}</b> {count, plural, =0 {entries} one {entry} other {entries}} selected',
-            },
-            {
-              count: entriesToPublish.length,
-              b: BoldChunk,
-            }
-          )}
-        </Typography>
-        <Table.Root rows={entries} colCount={4}>
-          <Box marginTop={5}>
-            <SelectedEntriesTableContent notifySelectionChange={setEntriesToPublish} />
-          </Box>
-        </Table.Root>
-      </ModalBody>
-      <ModalFooter
-        startActions={
-          <Button onClick={onToggle} variant="tertiary">
-            {formatMessage({
-              id: 'app.components.Button.cancel',
-              defaultMessage: 'Cancel',
-            })}
-          </Button>
-        }
-        endActions={
-          <Button onClick={onConfirm} disabled={entriesToPublish.length === 0}>
-            {formatMessage({ id: 'app.utils.publish', defaultMessage: 'Publish' })}
-          </Button>
-        }
-      />
-    </ModalLayout>
-  );
-};
-
-SelectedEntriesModal.propTypes = {
-  isOpen: PropTypes.bool.isRequired,
-  onToggle: PropTypes.func.isRequired,
-  onConfirm: PropTypes.func.isRequired,
-};
 
 /* -------------------------------------------------------------------------------------------------
  * BulkActionButtons
